removed the template_sphere
