+++ OpenPNM/Network/__Cubic__.py
@@ -448,46 +448,3 @@ class Cubic(GenericNetwork):
     def subdivide(self, pores=[], shape=[], labels=[]):
         topo.subdivide(network=self, pores=pores, shape=shape, labels=labels)
     subdivide.__doc__ = topo.subdivide.__doc__
-
-    def template_sphere_shell(outer_radius=None, inner_radius=0):
-        r"""
-        This method generates an image array of a sphere shell for a cubic network.
-
-        Parameters
-        ----------
-        outer_radius : array_like
-        Number of the nodes in the outer radius of the shell
-
-        inner_radius : float
-        Number of the nodes in the inner radius of the shell
-
-        """
-
-        if outer_radius is None:
-            raise Exception('No outer radius has been sent!')
-        if inner_radius is None:
-            raise Exception('Number of nodes in the inner radius cannot be '
-                            'None!')
-        rmax = np.array(outer_radius, ndmin=1)
-        rmin = np.array(inner_radius, ndmin=1)
-        s_rmax = np.size(rmax)
-        s_rmin = np.size(rmin)
-        if not ((s_rmax in [1, 3]) and (s_rmin in [1, 3])):
-            raise Exception('In this method, each radius can be scalar or '
-                            'array with components along all xyz directions.')
-        s_u_rmax = np.size(np.unique(rmax))
-        s_u_rmin = np.size(np.unique(rmin))
-        if not ((s_u_rmax == 1) and (s_u_rmin == 1)):
-            raise Exception('In this method, all components of radius should '
-                            'be unique values along all xyz directions.')
-        pnum = 2 * np.ones(3) * rmax - 1
-        Rx, Ry, Rz = np.array(pnum, dtype=np.int32)
-        x, y, z = np.indices((Rx, Ry, Rz))
-        x = x - (Rx - 1)/2
-        y = y - (Ry - 1)/2
-        z = z - (Rz - 1)/2
-        img = x ** 2 + y ** 2 + z ** 2 < np.unique(rmax) ** 2
-        if not np.all(rmin == 0):
-            img_min = x ** 2 + y ** 2 + z ** 2 > np.unique(rmin) ** 2
-            img = img * img_min
-        return (img)
