+++ test/unit/Network/CubicTest.py
@@ -148,19 +148,3 @@ class CubicTest:
         L = self.net.domain_length(face_1=self.net.pores('top'),
                                    face_2=self.net.pores('bottom'))
         assert sp.allclose(L, 4, rtol=1e-02)
-
-    def test_template_sphere_shell(self):
-        from OpenPNM.Utilities import topology
-        spacing = sp.array([0.5])
-        r_o = [5, 5, 5]
-        r_in = [2, 2, 2]
-        img = OpenPNM.Network.Cubic.template_sphere_shell(outer_radius=r_o,
-                                                          inner_radius=r_in)
-        pn_sphere = OpenPNM.Network.Cubic(template=img, spacing=spacing)
-        assert pn_sphere.Np == 452
-        img2 = OpenPNM.Network.Cubic.template_sphere_shell(outer_radius=r_o)
-        pn_sphere = OpenPNM.Network.Cubic(template=img2, spacing=spacing)
-        L1 = sp.amax(topology.find_pores_distance(network=pn_sphere,
-                                                  pores1=pn_sphere.Ps,
-                                                  pores2=pn_sphere.Ps))
-        assert L1 < sp.sqrt(3) * 8 * 0.5
